Silicon/SynQuacer: set PHY mode as appropriate in ACPI and DT tables
As it turns out, the DeveloperBox platform never described its Ethernet
PHY mode correctly: the 'rgmii' value it exposes to the OS was inherited
from the SynQuacer evaluation board, which uses a different PHY, and the
Realtek PHY used on DeveloperBox is integrated on the board with straps
that configure it to 'rgmii-id' mode.

We never noticed because the Realtek PHY driver in Linux ignored the PHY
mode to begin with, and simply used the configuration that was active at
boot. Unfortunately, that has changed, and recent versions of the  Linux
kernel (including stable releases) will now honour the firmware provided
PHY mode, and therefore configure the PHY incorrectly on these boards,
resulting in loss of network connectivity.

For ACPI boot, we can fix this by just setting the PHY mode to the empty
string - the Linux driver will be updated (and the change backported) to
ignore it anyway, as ACPI boot implies rich firmware, and it is reasonable
to assume that the PHY will be configured before the OS boots.

For DT, let's fix the description instead. This involves moving the
'phy-mode' property out of the shared .dtsi, as the change should only
apply to DeveloperBox, not to the SynQuacer evaluation board.

Signed-off-by: Ard Biesheuvel <ard.biesheuvel@arm.com>
Reviewed-by: Leif Lindholm <leif@nuviainc.com>