Silicon/Qemu/Sbsa: sbsa-wdt interrupt id update
The previous value of interrupt id used was not in sync
with the interrupt id being used in qemu sbsa-ref platform
due to a conflict before merging and was missed in last review.
This was preventing the watchdog interrupt from getting
identified.Updated SBSA-wdt interrupt id in Gtdt table
to rectify the issue.

Cc: Leif Lindholm <leif@nuviainc.com>
Cc: Ard Biesheuvel <ard.biesheuvel@arm.com>
Cc: Graeme Gregory <graeme@nuviainc.com>
Signed-off-by: Shashi Mallela <shashi.mallela@linaro.org>
Reviewed-by: Leif Lindholm <leif@nuviainc.com>