Platform/AMD/Styx: implement PCIe _DSM #5 to preserve boot allocations
Now that we have gained support for resizable BARs in EDK2, there is no
longer a need for the OS to reallocate MMIO BAR resources. So let's
implement _DSM #5 properly to inform the OS that boot time allocations
may be preserved.

While at it, clean up the implementation of _DSM, and only retain the
methods that we actually need: #0 exposing the fact that #0 and #5 are
the only ones implemented, and #5 as explained above.

Signed-off-by: Ard Biesheuvel <ard.biesheuvel@arm.com>
Reviewed-by: Leif Lindholm <leif@nuviainc.com>