Platform/AMD/Styx: remove unused RP0 device
The RP0 device has no _HID or _CID or any other identifying attribute,
and its use of _ADR within the scope of the PNP0A08 device object suggests
that _ADR carries a PCIe dev/fn identifier (like the EXPn device objects
earlier in the file), whereas in reality, the address it carries is the
MMIO address of the ECAM window.

So let's just drop it.

Signed-off-by: Ard Biesheuvel <ard.biesheuvel@arm.com>
Reviewed-by: Leif Lindholm <leif@nuviainc.com>