Platform/ARM: FVP: Fix serial port usage
The serial ports on FVP are follows:
  UART#   Address        Interrupt No.
  -----   -------        -------------
  UART0   0x1c090000     0x25
  UART1   0x1c0a0000     0x26
  UART2   0x1c0b0000     0x27
  UART3   0x1c0c0000     0x28

See Fast Model Reference Manual at:
https://developer.arm.com/documentation/100964/1113/

The FVP firmware configures the debug log output on UART1 and the
UEFI shell is configured to use UART0. The Serial Port Console
Redirector (SPCR) table is configured such that the UEFI shell is
migrated to the OS serial console.

When Dynamic Tables Framework is enabled, the FVP Configuration
Manager uses the PCDs PcdSerialRegisterBase and PL011UartInterrupt
to populate the SPCR table.
The current firmware implementation has two problems:
 1. The PL011UartInterrupt was not set to the correct interrupt
    number. Therefore, the OS serial console was not fully
    functional (serial output was printed but input could not
    be provided).
 2. The OS serial console was setup on UART1, while it should
    have been UART0.

Therefore, when Dynamic Tables Framework is enabled fix the UART
base address and interrupt number such that UART0 is configured
in the SPCR table. Also remove ConfigurationManager.dsc.inc as
this is now redundant.

Signed-off-by: Sami Mujawar <sami.mujawar@arm.com>
Reviewed-by: Leif Lindholm <leif@nuviainc.com>