Platform/ARM: FVP: Remove Serial port description from DSDT
Dynamic Tables Framework has been updated to add support for
Dynamic AML. The SPCR and DBG generators have also been updated
to dynamically generate the AML description for the Serial Ports.
Therefore, the serial port description in the DSDT is redundant
when Dynamic Tables Framework is enabled, so remove it.

Signed-off-by: Sami Mujawar <sami.mujawar@arm.com>
Reviewed-by: Leif Lindholm <leif@nuviainc.com>