Platform/Arm: Fix Ecc error 7008 for SCMI_CLOCK_RATE
Following the Ecc reported error in the edk2 repository,
the SCMI_CLOCK_RATE structure has been modified in:
an enum and its elements have been renamed in:
ArmPkg/Include/Protocol/ArmScmiClockProtocol.h

This patch is a follow-up and fixes the following Ecc
reported error:
Complex types should be typedef-ed

The error is due to the a nested structure declaration.

The patch also re-formats the debug messages to fit in
a 80 chars line.

Signed-off-by: Pierre Gondois <Pierre.Gondois@arm.com>
Reviewed-by: Sami Mujawar <sami.mujawar@arm.com>