Marvell/Cn9132Db: Enable ACPI support
Because of the limitation in a number of available intterrupts
that can be mapped in a current version of the IcuLib,
the ACPI support for Cn9132 variant remained disabled.
Such hard limitation is not needed though and enable ACPI
boot, however with a the interfaces present only on the
first two CP115 HW blocks.

Signed-off-by: Marcin Wojtas <mw@semihalf.com>
Reviewed-by: Ard Biesheuvel <ardb@kernel.org>