Marvell: Armada7k8kPciHostBridgeLib: Remove ECAM base limitation
On CN913x-based platforms it is possible to have up to 9 PCIE
root complexes. In such case it may be necessary to configure
more configuration spaces with smaller bus count, so that
to fit the memory layout constraints. For that purpose remove
forcing ECAM base to be divisible by SIZE_256MB.

Signed-off-by: Marcin Wojtas <mw@semihalf.com>