KabylakeOpenBoardPkg/AspireVn7Dash572G: Fix Visual Studio Build
AspireVn7Dash572G currently does not build with Visual Studio.
This is due to the Visual C++ compiler generating warnings with the GCC
compiler does not. The two classes of issues are unused local variables
and implicit integer casts that could result in truncation. Visual C++
requires an explicit cast in cases where integer truncation is possible.

Cc: Chasel Chiu <chasel.chiu@intel.com>
Cc: Benjamin Doron <benjamin.doron00@gmail.com>
Signed-off-by: Nate DeSimone <nathaniel.l.desimone@intel.com>
Reviewed-by: Chasel Chiu <chasel.chiu@intel.com>